refactor(integration-tests): split plugin_dispatch.rs into focused submodules
Converts the 1585-line monolithic test file into a module with 10
focused subfiles, all under 300 lines each.

- fixtures.rs       — shared plugin types (TestPlugin, EchoTool, etc.)
- helpers.rs        — runtime builder helpers
- dispatch_tests.rs — core dispatch, not-found, registry builder
- security_tests.rs — security interceptor denial, workspace boundary
- error_multi_tests.rs — error propagation, multi-plugin dispatch
- hook_tests.rs     — PreToolCall, PostToolCall, ToolError hooks
- hot_reload_tests.rs  — plugin unloaded/failed mid-turn race conditions
- audit_tests.rs    — audit log entry verification
- tool_name_tests.rs   — special characters, colon-containing names
- kv_store_tests.rs    — KV store session isolation and cleanup

Uses #[path] attributes in the entry point since the crate root
resolves mod declarations relative to tests/, not tests/plugin_dispatch/.
